Update "Map Special Modifiers"
Map special modifiers now uses Prestigious when prestiging, uses the best metal cache source when farming for equipment, and uses Fast Attacks in all other situations (such as farming map stacks). In all cases it will use the best modifier that can be afforded. This is *not* retroactive when the automapper wants to use a map level that already exists.